+++ src/Spreads.Collections/Collections/SortedDeque.fs
@@ -1,4 +1,4 @@
-﻿namespace Spreads.Collections.Experimental
+﻿namespace Spreads.Collections
 
 open System
 open System.Collections
@@ -10,59 +10,111 @@ open System.Runtime.InteropServices
 open Spreads
 open Spreads.Collections
 
+type KVComparer<'K,'V when 'K : comparison>(keyComparer:IComparer<'K>, valueComparer:IComparer<'V>) = 
+  interface IComparer<KV<'K,'V>> with
+    member this.Compare(x: KV<'K, 'V>, y: KV<'K, 'V>): int = 
+      let c1 = keyComparer.Compare(x.Key, y.Key)
+      if c1 = 0 then valueComparer.Compare(x.Value, y.Value)
+      else c1
 
-/// A shortcut for KeyValuePair
-//[<CustomComparison;CustomEquality>]
-//type KV<'K,'V when 'K : comparison and 'K : equality> =
-//  struct
-//    val Key : 'K
-//    val Value : 'V
-//    new(key, value) = {Key = key; Value = value}
-//  end
-//  override x.Equals(yobj) =
-//    match yobj with
-//    | :? KV<'K,_> as y -> (x.Key = y.Key)
-//    | _ -> false
-//  override x.GetHashCode() = x.Key.GetHashCode()
-//  interface System.IComparable<KVP<'K,'V>> with
-//    member x.CompareTo y = 
-//      let c = Comparer<'K>.Default
-//      c.Compare(x.Key, y.Key)
-//  interface System.IComparable with
-//    member x.CompareTo other = 
-//      match other with
-//      | :? KV<'K,_> as y -> 
-//        let c = Comparer<'K>.Default
-//        c.Compare(x.Key, y.Key)
-//      | _ -> invalidArg "other" "Cannot compare values of different types"
+and
+  // A shortcut for KeyValuePair
+  [<CustomComparison;CustomEquality>]
+  KV<'K,'V when 'K : comparison> =
+    struct
+      val Key : 'K
+      val Value : 'V
+      new(key, value) = {Key = key; Value = value}
+    end
+    override x.Equals(yobj) =
+      match yobj with
+      | :? KV<'K,_> as y -> (x.Key = y.Key)
+      | _ -> false
+    override x.GetHashCode() = x.Key.GetHashCode()
+    interface System.IComparable<KV<'K,'V>> with
+      member x.CompareTo y = 
+        let c1 = Comparer<'K>.Default.Compare(x.Key, y.Key)
+        if c1 = 0 then 
+          Comparer<'V>.Default.Compare(x.Value, y.Value)
+        else c1
+    interface System.IComparable with
+      member x.CompareTo other = 
+        match other with
+        | :? KV<'K,'V> as y -> 
+          (x :> System.IComparable<KV<'K,'V>>).CompareTo(y)
+        | _ -> invalidArg "other" "Cannot compare values of different types"
 
 
 
 [<SerializableAttribute>]
-type SortedDeque<'T when 'T : comparison and 'T : equality>
+type SimpleSortedDeque<'T when 'T : comparison>
+  internal(comparer:IComparer<'T>) as this=
+  [<DefaultValue>] val mutable comparer : IComparer<'T> 
+  [<DefaultValue>] val mutable list : List<'T>
+
+  do
+    this.comparer <- if comparer = null then Comparer<'T>.Default :> IComparer<'T> else comparer
+    this.list <- List<'T>()
+
+  member this.Add(item:'T) = 
+    if this.list.Count = 0 || this.comparer.Compare(item, this.list.[this.list.Count - 1]) > 0 then
+      this.list.Add(item)
+    else
+      let idx = this.list.BinarySearch(item, this.comparer)
+      if idx >= 0 then invalidOp "Item already exists"
+      else this.list.Insert(~~~idx, item)
+
+  member this.First with get() = this.list.[0]
+  member this.Last with get() = this.list.[this.list.Count - 1]
+  member this.Count with get() = this.list.Count
+
+  member this.RemoveFirst() : 'T = 
+    let first = this.First
+    this.list.RemoveAt(0)
+    first
+  
+  interface IEnumerable with
+    member this.GetEnumerator() = this.list.GetEnumerator() :> IEnumerator
+  interface IEnumerable<'T> with
+    member this.GetEnumerator() = this.list.GetEnumerator() :> IEnumerator<'T>
+
+
+
+
+[<SerializableAttribute>]
+[<ObsoleteAttribute("This is fucking manifest of complexity. Was badly rewritten 2+ years ago from not very good source")>]
+type SortedDeque<'T when 'T : comparison>
   internal(input:IEnumerable<'T>, capacity1:int, comparer:IComparer<'T>) as this=
-  let mutable cpct : int = capacity1
+  let mutable capacity : int = capacity1
   [<DefaultValue>] val mutable comparer : IComparer<'T> 
   [<DefaultValue>] val mutable buffer : 'T array 
   [<DefaultValue>] val mutable firstOffset : int
   [<DefaultValue>] val mutable size : int
   [<DefaultValue>] val mutable version : int
-  [<DefaultValue>] val mutable isReadOnly : bool
-  [<DefaultValue>] val mutable maxSize : int
-  //[<DefaultValue>] val mutable syncObject : obj
   [<DefaultValue>] val mutable isSynchronized : bool
   [<NonSerializedAttribute>]
   let syncRoot = new Object()
   do
     this.comparer <- if comparer = null then LanguagePrimitives.FastGenericComparer else comparer
     this.buffer <- Array.zeroCreate capacity1
-    this.maxSize <- 2146435071
     this.Capacity <- capacity1
     
   //#region Private & Internal members
 
   /// Calculates offset from index
-  member inline private this.IndexToOffset(index) = (index + this.firstOffset) % cpct
+  member inline private this.IndexToOffset(index) = (index + this.firstOffset) % capacity
+  member private this.OffsetToIndex(offset) = 
+    // TODO this is fucking mess, the whole structure is such a mess! 
+    let offset' = if offset < 0 then ~~~ offset else offset
+    let ret = offset' - this.firstOffset
+    let res = 
+      if ret > 0 then ret
+      else
+  //    if offset < 0 then 
+  //      ~~~((~~~offset - this.firstOffset + capacity) % capacity)
+  //    else
+        (offset' - this.firstOffset + capacity) % capacity
+    if offset < 0 then ~~~res else res
 
   member inline private this.GetByIndex(index) : 'T =
     if index <= this.size then
@@ -85,7 +137,7 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
 
   member inline private this.AddToLeft(element) =
     this.EnsureCapacity()
-    let offset = (this.firstOffset - 1 + cpct) % cpct
+    let offset = (this.firstOffset - 1 + capacity) % capacity
     this.firstOffset <- offset
     this.buffer.[offset] <- element
     this.size <- this.size + 1
@@ -103,16 +155,16 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
       // TODO check for split and do array.copy when possible
       if index < this.size / 2 then
           let copyCount = index
-          let writeIndex = cpct - 1
+          let writeIndex = capacity - 1
           for j in 0..copyCount-1 do
               this.buffer.[this.IndexToOffset(writeIndex + j)] 
                 <- this.buffer.[this.IndexToOffset(j)]
-          let offset = (this.firstOffset - 1 + cpct) % cpct
+          let offset = (this.firstOffset - 1 + capacity) % capacity
           this.firstOffset <- offset
       else
           let copyCount = this.size - index
           let writeIndex = index + 1
-          for j in copyCount-1..0 do
+          for j in (copyCount-1)..(-1)..0 do
               this.buffer.[this.IndexToOffset(writeIndex + j)] <- this.buffer.[this.IndexToOffset(index + j)]
       let offset = this.IndexToOffset(index)
       this.buffer.[offset] <- element
@@ -131,7 +183,7 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
     if this.size > 0 then
       this.size <- this.size - 1
       let offset = this.firstOffset
-      this.firstOffset <- (this.firstOffset + 1) % cpct
+      this.firstOffset <- (this.firstOffset + 1) % capacity
       this.version <- this.version + 1
       this.buffer.[offset]
     else raise (InvalidOperationException("Deque is empty"))
@@ -139,7 +191,7 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
   member inline private this.RemoveAtIndex(index) : 'T =
     let ret = this.GetByIndex(index)
     if index = 0 then
-        this.firstOffset <- (this.firstOffset + 1) % cpct
+        this.firstOffset <- (this.firstOffset + 1) % capacity
     elif index = this.size - 1 then
         // decrement size at the end of the method
         ()
@@ -148,7 +200,7 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
             let copyCount = index
             for j in copyCount-1..0 do
                 this.buffer.[this.IndexToOffset(j + 1)] <- this.buffer.[this.IndexToOffset(j)]
-            this.firstOffset <- (this.firstOffset + 1) % cpct
+            this.firstOffset <- (this.firstOffset + 1) % capacity
         else
             let copyCount = this.size - index - 1
             let readIndex = index + 1
@@ -162,7 +214,7 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
     Trace.Assert(this.size <= this.buffer.Length)
     if this.size = this.buffer.Length then 
       let mutable num = ((this.buffer.Length + 1) * 3) / 2
-      if num > this.maxSize then num <- this.maxSize
+      if num > 2146435071 then num <- 2146435071
       if min.IsSome && num < min.Value then num <- min.Value
       this.Capacity <- num
     
@@ -171,7 +223,7 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
   /// Removes first element to free space for new element if the map is full. 
   member internal this.AddAndRemoveFisrtIfFull(element) =
       //use lock = makeLock this.SyncRoot
-      if this.size = cpct then this.RemoveFromLeft() |> ignore
+      if this.size = capacity then this.RemoveFromLeft() |> ignore
       this.Add(element)
       ()
 
@@ -181,7 +233,7 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
 
   ///Gets or sets the capacity. This value must always be greater than zero, and this property cannot be set to a value less than this.size/>.
   member this.Capacity 
-      with get() = cpct
+      with get() = capacity
       and set(value) =
         let entered = enterLockIf syncRoot  this.isSynchronized
         try
@@ -190,8 +242,8 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
           | c when c < this.size -> raise (ArgumentOutOfRangeException("Small capacity"))
           | c when c > 0 -> 
               let elArr : 'T array = Array.zeroCreate c
-              if this.firstOffset + this.size > cpct then // is split
-                  let len = cpct - this.firstOffset
+              if this.firstOffset + this.size > capacity then // is split
+                  let len = capacity - this.firstOffset
                   // Elements
                   Array.Copy(this.buffer, this.firstOffset, elArr, 0, len);
                   Array.Copy(this.buffer, 0, elArr, len, this.size - len);
@@ -200,7 +252,7 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
               this.buffer <- elArr
               this.firstOffset <- 0
               //this.version <- this.version + 1
-              cpct <- value
+              capacity <- value
           | _ -> ()
         finally
           exitLockIf syncRoot entered
@@ -271,8 +323,8 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
       with get()=
         let entered = enterLockIf syncRoot  this.isSynchronized
         try
-          if this.size = 0 then OptionalValue.Missing
-          else OptionalValue(this.GetByIndex(0)) 
+          if this.size = 0 then invalidOp "Deque is empty"
+          else this.GetByIndex(0) 
         finally
           exitLockIf syncRoot entered
 
@@ -280,12 +332,12 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
       with get() =
         let entered = enterLockIf syncRoot  this.isSynchronized
         try
-          if this.size = 0 then OptionalValue.Missing
-          else OptionalValue(this.GetByIndex(this.size - 1)) 
+          if this.size = 0 then invalidOp "Deque is empty"
+          else this.GetByIndex(this.size - 1) 
         finally
           exitLockIf syncRoot entered
 
-  member this.Contains(element:'T) = this.IndexOfElement(element) >= 0
+  member this.Contains(element:'T) = this.OffsetOfElement(element) >= 0
 
   member this.Contains(predicate:'T -> bool) = this.IndexOfFirst(predicate) >= 0
 
@@ -344,7 +396,7 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
       if obj.Equals(element, null) then raise (ArgumentNullException("element"))
       //use lock = makeLock this.SyncRoot
       // TODO Last/Fisrt optimization here
-      let index = this.IndexOfElement(element)
+      let index = this.OffsetToIndex(this.OffsetOfElement(element))
       if index >= 0 then raise (ArgumentException("key already exists"))
       try
         let target = ~~~index
@@ -373,21 +425,23 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
     finally
       exitLockIf syncRoot entered
 
-  member this.IndexOfElement(element:'T) =
+  member this.IndexOfElement(element:'T) = this.OffsetToIndex(this.OffsetOfElement(element))
+
+  member private this.OffsetOfElement(element:'T) =
     let entered = enterLockIf syncRoot  this.isSynchronized
     try
       if obj.Equals(element, null) then raise (ArgumentNullException("element"))
       let mutable index = 0
-      if this.firstOffset + this.size > cpct then // is split
+      if this.firstOffset + this.size > capacity then // is split
           let c = this.Comparer.Compare(element, this.buffer.[0]) // changed from this.GetByIndex(0) // this.buffer.[0]
           match c with
-          | 0 -> index <- this.Capacity - this.firstOffset
+          | 0 -> index <- 0 //this.Capacity - this.firstOffset
           | -1 -> // key in the right part of the buffer
               index <- 
-                Array.BinarySearch(this.buffer, this.firstOffset, cpct - this.firstOffset, element, this.comparer) 
+                Array.BinarySearch(this.buffer, this.firstOffset, capacity - this.firstOffset, element, this.comparer) 
           | 1 -> // key in the first part of the buffer
               index <- 
-                Array.BinarySearch(this.buffer, 0, this.firstOffset - (cpct - this.size), element, this.comparer)
+                Array.BinarySearch(this.buffer, 0, this.firstOffset - (capacity - this.size), element, this.comparer)
           | _ -> failwith("nonsense")
       else
           index <- Array.BinarySearch(this.buffer, this.firstOffset, this.size, element, this.comparer) 
@@ -395,33 +449,33 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
     finally
       exitLockIf syncRoot entered
 
-  member this.Item
-      with get element =
-        let entered = enterLockIf syncRoot  this.isSynchronized
-        try
-          let index = this.IndexOfElement(element)
-          if index >= 0 then
-              this.GetByIndex(index)
-          else
-              raise (KeyNotFoundException())
-        finally
-          exitLockIf syncRoot entered
-      and set oldElement newElement =
-        if obj.Equals(oldElement, null) then raise (ArgumentNullException("oldElement"))
-        let entered = enterLockIf syncRoot  this.isSynchronized
-        try
-          let index = this.IndexOfElement(oldElement)
-          if index >= 0 then // contains key            
-              this.SetByIndex(index, newElement)
-          else
-              this.InsertAtIndex(~~~index, newElement)
-        finally
-          exitLockIf syncRoot entered
+//  member this.Item
+//      with get element =
+//        let entered = enterLockIf syncRoot  this.isSynchronized
+//        try
+//          let index = this.IndexOfElement(element)
+//          if index >= 0 then
+//              this.GetByIndex(index)
+//          else
+//              raise (KeyNotFoundException())
+//        finally
+//          exitLockIf syncRoot entered
+//      and set oldElement newElement =
+//        if obj.Equals(oldElement, null) then raise (ArgumentNullException("oldElement"))
+//        let entered = enterLockIf syncRoot  this.isSynchronized
+//        try
+//          let index = this.IndexOfElement(oldElement)
+//          if index >= 0 then // contains key            
+//              this.SetByIndex(index, newElement)
+//          else
+//              this.InsertAtIndex(~~~index, newElement)
+//        finally
+//          exitLockIf syncRoot entered
 
   member this.Remove(element, [<Out>]result: byref<'T>) : bool =
     let entered = enterLockIf syncRoot  this.isSynchronized
     try
-      let index = this.IndexOfElement(element)
+      let index = this.OffsetToIndex(this.OffsetOfElement(element))
       if index >= 0 then this.RemoveAtIndex(index) |> ignore
       index >= 0
     finally
@@ -449,7 +503,7 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
     let idx = ref -1
     let entered = enterLockIf syncRoot  this.isSynchronized
     try
-      let index = this.IndexOfElement(element)
+      let index = this.OffsetToIndex(this.OffsetOfElement(element))
       match lookup with
       | Lookup.EQ -> 
           if index >= 0 then
@@ -520,9 +574,9 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
   member this.TryGet(element, [<Out>]value: byref<'T>) : bool =
     let entered = enterLockIf syncRoot  this.isSynchronized
     try
-      let index = this.IndexOfElement(element)
-      if index >= 0 then
-          value <- this.GetByIndex(index)
+      let offset = this.OffsetOfElement(element)
+      if offset >= 0 then
+          value <- this.buffer.[offset]
           true
       else
           false
@@ -539,8 +593,8 @@ type SortedDeque<'T when 'T : comparison and 'T : equality>
     
   //#region Constructors
 
-  new() = SortedDeque([||], 8, Comparer<'T>.Default)
-  new(comparer:IComparer<'T>) = SortedDeque([||], 8, comparer)
+  new() = SortedDeque([||], 2, Comparer<'T>.Default)
+  new(comparer:IComparer<'T>) = SortedDeque([||], 2, comparer)
   new(capacity:int) = SortedDeque([||], capacity, Comparer<'T>.Default)
   new(capacity:int,comparer:IComparer<'T>) = SortedDeque([||], capacity, comparer)