+++ src/Spreads.Collections/Spreads.Series.2.fs
@@ -5,15 +5,20 @@ open System.Collections
 open System.Collections.Generic
 open System.Diagnostics
 open System.Runtime.InteropServices
-
+open System.Runtime.CompilerServices
 open Spreads
 
 // could extend but not operators
+[<AutoOpenAttribute>]
 module SpreadsModule =
-  type BaseSeries with
-    static member private Init() =
-      ()
+  type Series with
+    static member private Init() = ()
+
+    static member Zip<'K,'V,'R when 'K : comparison>(resultSelector:Func<'K,'V[],'R>,[<ParamArray>] series: Series<'K,'V> array) =
+      CursorSeries(fun _ -> new ZipNCursor<'K,'V,'R>(resultSelector, series |> Array.map (fun s -> s.GetCursor))  :> ICursor<'K,'R>) :> Series<'K,'R>
+
     end
+  
 
 module internal Initializer =
   let internal init() =